requirements.txt
absl-py==2.1.0
accelerate==0.21.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.3.0
backoff==2.2.1
bitsandbytes==0.41.0
blinker==1.9.0
Brotli==1.0.9
certifi==2025.1.31
charset-normalizer==3.3.2
click==8.1.8
cmake==4.0.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.2.2
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.13.1
Flask==3.1.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
gensim==4.3.3
gmpy2==2.2.1
googletrans==4.0.2
gradio==3.35.2
gradio_client==0.2.9
grpcio==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.29.2
hyperframe==6.1.0
idna==3.7
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
latex2mathml==3.77.0
linkify-it-py==2.0.3
lit==18.1.8
lxml==5.3.1
Markdown==3.7
markdown-it-py==2.2.0
markdown2==2.5.3
MarkupSafe==2.1.5
matplotlib==3.9.4
mdit-py-plugins==0.3.3
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft==1.3.11
mkl_random==1.2.8
mpmath==1.3.0
multidict==6.2.0
narwhals==1.32.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
openai==1.12.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
peft==0.4.0
pillow==10.4.0
portalocker==3.1.1
propcache==0.3.1
protobuf==6.30.0
psutil==7.0.0
pydantic==1.10.21
pydantic_core==2.33.0
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
rich==13.9.4
rouge_score==0.1.2
rpds-py==0.24.0
ruff==0.11.2
sacrebleu==2.5.1
safetensors==0.5.3
scikit-learn==1.2.2
scipy==1.13.1
semantic-version==2.10.0
sentencepiece==0.1.99
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
starlette==0.46.1
svgwrite==1.4.3
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
textaugment==2.0.0
textblob==0.19.0
threadpoolctl==3.5.0
tiktoken==0.9.0
timm==0.9.12
tokenizers==0.15.1
tomlkit==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.67.1
transformers==4.36.2
triton==2.0.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
urllib3==2.3.0
wavedrom==2.0.3.post3
websockets==11.0.3
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3

